world speed being float value

+++ pygreenfoot/world.py
@@ -36,7 +36,7 @@ class World(metaclass=ABCMeta):
         self.__app = Application.get_app()
         self.__texts: Dict[Tuple[int, int], Tuple[Text, Tuple[int, int]]] = {}
         self.running: bool = True
-        self.world_speed: int = self.__app.default_world_speed
+        self.world_speed: float = self.__app.default_world_speed
         self.__last_time = time()
         
     def add_to_world(self, game_object: "Actor", x: int, y: int) -> None:
